refactor: folder structure
move into modules/
