+++ src/app/(auth)/chat/page.tsx
@@ -1,4 +1,4 @@
-import Chat from "@/components/chat";
+import Chat from "@/app/modules/chat/ui/components/chat";
 import { Suspense } from "react";
 
 export default function ChatPage() {
