+++ src/app/modules/map/lib/create-custom-overlay.ts
